+++ src/main/java/io/mailtrap/client/MailtrapClient.java
@@ -74,7 +74,7 @@ public class MailtrapClient {
      *
      * @param inboxId the inbox ID to use for testing
      */
-    public void switchToEmailTestingApi(Integer inboxId) {
+    public void switchToEmailTestingApi(Long inboxId) {
         this.sendingContextHolder.setInboxId(inboxId);
         this.sendingContextHolder.setSandbox(true);
         this.sendingContextHolder.setBulk(false);
